Rename reactor module to mockreactor. Use MockReactor instead of MemoryReactor.
